backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
spacy==3.7.2
joblib==1.3.2
psycopg2-binary==2.9.9
